+++ test_app/lib/test_app.rb
@@ -105,5 +105,6 @@ module TestApp
   end
 end
 
+require 'test_app/metaclasses'
 require 'test_app/literal'
 $VERBOSE = original